refactor: add domain package
- Move Car, Racing class to domain package
