+++ src/main/java/racingcar/view/Output.java
@@ -1,6 +1,6 @@
 package racingcar.view;
 
-import racingcar.Car;
+import racingcar.domain.Car;
 import java.util.ArrayList;
 
 public class Output {
