+++ src/test/java/racingcar/CarTest.java
@@ -5,6 +5,8 @@ import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import racingcar.domain.Car;
+
 import static org.assertj.core.api.Assertions.assertThat;
 import java.util.ArrayList;
